Add coc and extensions / remove what it replaces
I've commented out some things and deleted others. This is inconsistent
and the reason that I'm not reaproaching this in a more organised way is
simply because this is a very uncertain process for me which has taken a
long time to find stability. I really don't want to risk making any more
changes and finding that I've reintroduced something heavy or disruptive
and struggle to get back to this place.

It's not all that bad and I'll attempt some tidying in the next few
commits with confidence that I can get back here if needed.