Configure vim command to alias to nvim
I suppose this is because my fingers naturally type the vim so often,
but it's also in case there are processes which are configured to open
up vim. I'm thinking os processes. This would probably be a matter for
reconfiguration, but since I use this in the office and don't have the
ability to rapidly develop fixes like this, I prefer to just have the
alias in place to avoid issues.