Add telescope plugin
I've included a couple of supporting plugins which I'm not too sure of
the function, but I'll come back to them. The devicons plugin is optional
and won't work well until I've installed a patched font, but I'm going
to add it and experiment with patched fonts in it's presence.

There are plenty of other key bindings I'll want to play with here too
but these will get us off to a start.