Return treesitter to installing all parsers
This is being changed about a lot in this branch. There are 2 reasons:

1. I don't want to be running test builds with this set to `all` because
then I have to wait for them all to install each time I test. Sometimes
I accidentally leave the modification to `javascript` or something as
short in there for the commit.

2. I can't actually decide whether I want this config to differ from the
Dockerfile's manual install of the parsers. Something tells me that I
should keep this actual file slimmer than the Dockerfile because the
Dockerfile will be built unattended, while if I actually use this file
outside of the Docker image produced by
https://github.com/js-jslog/development-env then I'll not want to be
waiting around for them all.

None of this really matters though so I'll just leave it at that.